requirements.txt
opencv-python
pyzbar
qrcode[pil]
Pillow
matplotlib
pandas
openpyxl
